Better `long` and `mpq_class` support, and more consistent usage of functors
Another step towards a "simple" low level interface for languages using only long and double types that uses mpz_class for internal computations but can accept a variety of "user facing" types. E.g. for matlab me might want to have an interface of the form

  double -> ( internal computations ) -> double

In the case of the volume, this could be expanded to something like

    double -> mpq_class -> ( internal computations ) -> mpq_class -> double,

where the double <-> mpq_class conversions are not yet implemented. Unlike the volume, the output halfspaces can always be guaranteed to be integral for any rational input generators, and, as of this commit, we have an (untested) interface of the form

    double -> mpq_class -> ( internal computations ) -> long

In this case, an (unsafe) conversion long -> double is easy to implement on demand.