simplified event points and abstracted Vector_t
The simplification of the event points makes more useful in specialized
constructions like 3D.

The pulling the vector type out to a template is a step in allowing the
dimension to be fixed at compile time. Then, one could for example use
Vector_t = std::array<3, Number_t> in three dimensions and hope that the
compiler uses the extra knowledge to unroll loops where appropriate.